@@ -2,62 +2,61 @@
  * Entry point of the Election app.
  */
 import * as path from 'path';
-import * as url from 'url';
+import * as nodeEnv from '_utils/node-env';
 // eslint-disable-next-line import/no-extraneous-dependencies
-import { BrowserWindow, app } from 'electron';
+import { BrowserWindow, app, ipcMain } from 'electron';
 
-let mainWindow: Electron.BrowserWindow | null;
+let mainWindow: Electron.BrowserWindow | undefined;
 
-function createWindow(): void {
+function createWindow() {
   // Create the browser window.
   mainWindow = new BrowserWindow({
     height: 600,
     width: 800,
     webPreferences: {
-      webSecurity: false,
-      devTools: process.env.NODE_ENV !== 'production',
+      devTools: nodeEnv.dev,
+      preload: path.join(__dirname, './preload.bundle.js'),
+      webSecurity: nodeEnv.prod,
     },
   });
 
   // and load the index.html of the app.
-  mainWindow.loadURL(
-    url.format({
-      pathname: path.join(__dirname, './index.html'),
-      protocol: 'file:',
-      slashes: true,
-    }),
-  ).finally(() => { /* no action */ });
+  mainWindow.loadFile('index.html').finally(() => { /* no action */ });
 
   // Emitted when the window is closed.
   mainWindow.on('closed', () => {
     // Dereference the window object, usually you would store windows
     // in an array if your app supports multi windows, this is the time
     // when you should delete the corresponding element.
-    mainWindow = null;
+    mainWindow = undefined;
   });
 }
 
 // This method will be called when Electron has finished
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
-app.on('ready', createWindow);
+app.whenReady().then(() => {
+  if (nodeEnv.dev || nodeEnv.prod) createWindow();
 
-// Quit when all windows are closed.
+  app.on('activate', () => {
+    if (BrowserWindow.getAllWindows.length === 0) createWindow();
+  });
+}).finally(() => { /* no action */ });
+
+// Quit when all windows are closed, except on macOS. There, it's common
+// for applications and their menu bar to stay active until the user quits
+// explicitly with Cmd + Q.
 app.on('window-all-closed', () => {
-  // On OS X it is common for applications and their menu bar
-  // to stay active until the user quits explicitly with Cmd + Q
-  if (process.platform !== 'darwin') {
-    app.quit();
-  }
+  if (process.platform !== 'darwin') app.quit();
 });
 
-app.on('activate', () => {
-  // On OS X it"s common to re-create a window in the app when the
-  // dock icon is clicked and there are no other windows open.
-  if (mainWindow === null) {
-    createWindow();
-  }
+ipcMain.on('renderer-ready', () => {
+  // eslint-disable-next-line no-console
+  console.log('Renderer is ready.');
 });
 
 // In this file you can include the rest of your app"s specific main process
 // code. You can also put them in separate files and require them here.
+
+// eslint-disable-next-line import/prefer-default-export
+export const exportedForTests = nodeEnv.test ? { createWindow } : undefined;
